+++ changelog.md
@@ -1,8 +1,10 @@
 # 0.8.0
     - [debug] debug menu disabled by default
     - [debug] im/export also for progress (not only bookmarks)
+    - [debug] stats about cached conferences/events/recordings
     - [leanback] homescreen recommendations
     - [leanback] seek thumbnails from backend
+    - [leanback] row for recently updated conferences & events
     - [touch] complete overhaul of navigation
     - [touch] darkmode configurable
     - [touch] Global Search
@@ -10,7 +12,8 @@
     - [touch] navigation rework
     - [touch] new search/filtering for events
     - [touch] Streams view shows list of upcoming streamed events
-    - refactoring of details-page (play-handling, etc.)
+    - [common] refactoring of details-page (play-handling, etc.)
+    - [common] replace Casty dependency with own implementation
 
 # 0.7.2
     - [fix] completely remove Appcenter Distribution-Module to be compliant with Playstore policies (was only active for non production builds)